% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_mircount.R
\name{as.data.frame.mircount}
\alias{as.data.frame.mircount}
\title{as.data.frame.mircount}
\usage{
\method{as.data.frame}{mircount}(x)
}
\description{
as.data.frame method for \code{mircount} class.
}
\details{
It converts \code{raw_read_count} into a data.frame.
}
