% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_mircount.R
\name{as.matrix.mircount}
\alias{as.matrix.mircount}
\title{as.matrix.mircount}
\usage{
\method{as.matrix}{mircount}(x)
}
\description{
as.matrix method for \code{mircount} class.
}
\details{
It converts \code{raw_read_count} into a matrix
}
