% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_mircount.R
\name{as.mircount.default}
\alias{as.mircount.default}
\title{as.mircount.default}
\usage{
\method{as.mircount}{default}(
  raw_read_count,
  sample_library_sizes,
  genes_complete_annotation,
  working_gene_annot_var_name,
  genes,
  targets,
  sample_groups_var_name,
  total_species,
  files_processed,
  target_annotation_file_processed
)
}
\description{
Default method for \code{as.mircount}
}
