% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novel.R
\name{hairpinSet}
\alias{hairpinSet}
\title{hairpinSet}
\usage{
hairpinSet(dataLst, setType = "training")
}
\arguments{
\item{dataLst}{Data list produced by \code{\link{mirProcessML}}.}

\item{setType}{The type of the output \code{fasta} file. Options are \code{"training"} or \code{"test"}. Default is \code{"training"}.}
}
\value{
Outputs a \code{.fasta} file that can be used as the training set or test set for novel miRNA discovery. File names: training set \code{trainingSet.fasta}; test set \code{testSet.fasta}
}
\description{
Produce a \code{.fasta} file as the training or test hairpin data set from the raw read count files. This function only works under Unix or Unix-like operating systems, with \code{pcregrep} and \code{parallel} installed.
}
\details{
Although the function is optimized for parallel computing, working with large reference file might stil result in long running time or system freezing, depending on the hardware configureation (CPU cores and RAM). Be sure to follow the file name naming convetion for refrence files: training reference \code{trainingRef.fasta}; test reference \code{testRef.fasta}.
}
\examples{
\dontrun{
mylist <- mirProcessML() # produce data list
hairpinSet(mylist, setType = "training") # produce the hairpin training set
}
}
