% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_mircount.R
\name{length.mircount}
\alias{length.mircount}
\title{length.mircount}
\usage{
\method{length}{mircount}(x)
}
\description{
length method for \code{mircount} class.
}
\details{
length is the number of samples
}
