% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conserved.R
\name{mirNrm}
\alias{mirNrm}
\title{mirNrm}
\usage{
mirNrm(dfm, count_threshold = "none", wgt = FALSE)
}
\arguments{
\item{dfm}{Input dataframe.}

\item{count_threshold}{Read count threshold. No filtering will be applied when set \code{"none"}. Otherwise, a numeric number can be set as the minimum read count for filtering. DDefault is \code{"none"}.}

\item{wgt}{If or not to apply sample weight. Default is \code{FALSE}.}
}
\value{
Outputs a dataframe with voom normalization
}
\description{
data Voom normalization for plotting
}
\examples{
\dontrun{
simNrm <- mirnrm(simDfm, wgt = TRUE)
}
}
