% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conserved.R
\name{mirProcess}
\alias{mirProcess}
\title{mirProcess}
\usage{
mirProcess(
  path = getwd(),
  species = NULL,
  target.annot.file = NULL,
  sample_groups.var.name = NULL,
  database = "mirbase",
  raw.file.sep = "",
  parallelComputing = FALSE,
  clusterType = "FORK",
  verbose = TRUE
)
}
\arguments{
\item{path}{Path to raw files. Default is the system working directory.}

\item{species}{Species code, following the traditional abbreviated naming convention, e.g. "hsa", "mmu".}

\item{target.annot.file}{Annotation file describing filenames and targets, and should be in \code{csv} format.}

\item{sample_groups.var.name}{Sample group annotation variable name in the \code{target.annot.file}.}

\item{database}{MiRNA database, only for miRNA naming conventions. Currently the function only takes "mirbase".}

\item{raw.file.sep}{Raw read count file separators. Default is \code{""\"\"}, i.e. white space.}

\item{parallelComputing}{Whether to use parallel computing or not. Default is \code{TRUE}.}

\item{verbose}{Bool. If to display function messages, excluding warnings and errors.}

\item{cluterType}{clusterType Only set when \code{parallelComputing = TRUE}, the type for parallel cluster. Options are \code{"PSOCK"} (all operating systems) and \code{"FORK"} (macOS and Unix-like system only). Default is \code{"PSOCK"}.}
}
\value{
Outputs a \code{mir_count} with merged read counts from multiple files, with annotation. The \code{mir_count} object contains the following:

         \code{raw_read_count}

         \code{sample_library_sizes}

         \code{genes_complete_annotation}: Data frame that contains all the annotation information for the miRNAs.

         \code{genes}: The associated miRNA names. The use of "gene" here is in a generic sense.

         \code{targets}: Sample annotation matrix

         \code{sample_gourps}: factor object for sample groups annotation

         \code{miRNA_database}

         \code{selected_species}

         \code{total_species}

         \code{files_processed}
}
\description{
Data pre-processing for miRNA-seq read count files. Also see \code{\link{mirProcessML}}.
}
\details{
When \code{species} left as \code{NULL}, the function will use all the species detected in the raw read count files.
         The raw count files are usually in \code{.txt} format with "read count" and "mirna" columns. The read count files can be obtained using the shell program \code{mirna_processing}.

         For \code{target.annot.file}, the argument doesn't accept full file path. The function will only seek the file under working directory. So, the file should be placed under working directory.
}
\examples{
\dontrun{
readcountMerged <- mirProcess()
}
}
