% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_mircount.R
\name{[,mircount,ANY,ANY,ANY-method}
\alias{[,mircount,ANY,ANY,ANY-method}
\title{\code{mircount} substting method}
\usage{
\S4method{[}{mircount,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\description{
Automatic subsetting method for \code{mircount} class.
}
\details{
It links the column of \code{raw_read_count} and rows of \code{targets}.
}
\examples{
\dontrun{
a <- new("mircount")
b <- a[1:5, 1:5]  # leads to raw_read_count[1:5, 1:5] and targets[1:5, ]
}
}
