% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_mircount.R
\docType{class}
\name{mircount-class}
\alias{mircount-class}
\alias{mircount}
\title{mircount}
\description{
S4 class equivalent to the S3 class \code{mir_count}.
}
\details{
1. The class links the column of \code{raw_read_count}, rows of \code{targets}, \code{sample_groups} and \code{genes}.
         2. Subsetting in "1" does not subset \code{genes_complete_annotation} as this data.frame may contain replicates.
}
\section{Slots}{

\describe{
\item{\code{raw_read_count}}{data.frame or matrix. Raw miRNA read counts, with row: genes, column: samples.}

\item{\code{sample_library_sizes}}{numeric. Vector for sample library sizes.}

\item{\code{genes_complete_annotation}}{data.frame. Complete gene annotation.}

\item{\code{genes}}{str. Vector for gene names.}

\item{\code{targets}}{data.frame. Complete sample annotation data frame.}

\item{\code{sample_groups_var_name}}{str. Variable name for sample grouping in the sample annotation.}

\item{\code{sample_groups}}{factor. Primary sample group factor.}

\item{\code{miRNA_database}}{str. MiR database.}

\item{\code{selected_species}}{str. The miRBase species code for the selected species of interest.}

\item{\code{total_species}}{str. Total number of species (in miRBase species code).}

\item{\code{files_processed}}{str. Processed file(s) name containing raw reads.}

\item{\code{target_annotation_file_processed}}{str. Processed sample annotation file name.}
}}

