% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conserved.R
\name{mirDeepProcess}
\alias{mirDeepProcess}
\title{mirDeepProcess}
\usage{
mirDeepProcess(
  raw.read.file = NULL,
  raw.file.sep = "/t",
  rm.norm.cols = TRUE,
  working.gene.annot.var.name = "miRNA",
  target.annot.file = NULL,
  sample_groups.var.name = NULL,
  sampleid.var.name = NULL,
  rep.merge.method = c("none", "sum", "mean"),
  species = NULL,
  database = "mirbase",
  parallelComputing = FALSE,
  clusterType = "FORK",
  verbose = TRUE
)
}
\arguments{
\item{raw.read.file}{Read count csv file from miRDeep2.}

\item{raw.file.sep}{Raw read count file separators. Default is \code{"\t"}, i.e. tab.}

\item{rm.norm.cols}{Bool. If to remove the \code{"(norm)"} columns from the read file. Default is \code{TRUE}.}

\item{working.gene.annot.var.name}{String. The column name for the miRNA identifiers. Default is \code{"miRNA"}. Could be precursor (\code{"precursor"}) as well.}

\item{target.annot.file}{Annotation file describing filenames and targets, and should be in \code{csv} format.}

\item{sample_groups.var.name}{Sample group annotation variable name in the \code{target.annot.file}.}

\item{sampleid.var.name}{Sample ID annotation variable name in the \code{target.annot.file}.}

\item{rep.merge.method}{String. MiRNA replicate merging method. Default is \code{"none"}.}

\item{species}{Species code, following the traditional abbreviated naming convention, e.g. "hsa", "mmu".}

\item{database}{MiRNA database, only for miRNA naming conventions. Currently the function only takes "mirbase".}

\item{parallelComputing}{Whether to use parallel computing or not. Default is \code{TRUE}.}

\item{verbose}{Bool. If to display function messages, excluding warnings and errors.}

\item{cluterType}{clusterType Only set when \code{parallelComputing = TRUE}, the type for parallel cluster. Options are \code{"PSOCK"} (all operating systems) and \code{"FORK"} (macOS and Unix-like system only). Default is \code{"PSOCK"}.}
}
\value{
Outputs a \code{mir_count},  with annotation. The \code{mir_count} object contains the following:

         \code{raw_read_count}

         \code{sample_library_sizes}

         \code{genes_complete_annotation}: Data frame that contains all the annotation information for the miRNAs.

         \code{working_gene_annot_var_name}: variable name for the working miRNA ids in the genes_complete_annotation data frame

         \code{genes}: The associated miRNA names. The use of "gene" here is in a generic sense. might have been merged

         \code{targets}: Sample annotation matrix

         \code{sample_groups_var_name}

         \code{sample_gourps}: factor object for sample groups annotation

         \code{miRNA_database}

         \code{selected_species}

         \code{total_species}

         \code{files_processed}

         \code{target_annotation_file_processed}: target.annot.file name
}
\description{
Data pre-processing for miRNA-seq read count files specifically designed for count results from the miRDeep2 conserved miRNA module.
}
\details{
1. When \code{species} left as \code{NULL}, the function will use all the species detected in the raw read count files.
            The raw count files are usually in \code{.csv} from miRDeep2. This file usually contains a \code{"#"} symbol in the first line, which is automatically removed by the current function.
            The first four columnes of the files are \code{miRNA}, \code{read_count}, \code{precursor} and \code{total}, forming the basis of \code{genes_complete_annotation} and \code{genes} in the output.

         2. For \code{target.annot.file}, the argument doesn't accept full file path. The function will only seek the file under working directory. So, the file should be placed under working directory.

         3. When miRNA replicates are merged, i.e. \code{rep.merge.method} set to either \code{"sum"} or \code{"mean"},
            the \code{genes} item contains the merged results, whereas the \code{genes_complete_annoation} contains the original first four columns
            from the miRDeep2 output, not merged.

            The replicates may happen when different precursors lead to the same mature miRNA.
}
\examples{
\dontrun{
readcountMerged <- mirDeepProcess(raw.read.file = "miRNAs_expressed_all_samples.csv", rm.norm.cols = TRUE,
                                  raw.file.sep = "/t",
                                  target.annot.file = "samples_annotation.csv",
                                  sample_groups.var.name = 'groupid',
                                  species = "mmu", database = "mirbase",
                                  parallelComputing = TRUE, clusterType = "FORK",
                                  verbose = TRUE)
}
}
