% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conserved.R
\name{mirFit}
\alias{mirFit}
\title{mirFit}
\usage{
mirFit(dfm, anno, fileName, wgt = FALSE)
}
\arguments{
\item{dfm}{Input dataframe.}

\item{anno}{Annoation file for the samples. Format is \code{csv} and make sure the second column is for the conditions.}

\item{fileName}{output file name. Be sure to use quatation}

\item{wgt}{If or not to apply sample weight. Default is \code{FALSE}.}
}
\value{
Outputs a list with limma eBayes fitting, QC plots and a volcano distribution
}
\description{
Linear fitting and emperical Bayesian statistical test, with the capability of producing volcano distribution, among other useful plots.
}
\examples{
\dontrun{
fitRNA <- mirFit(simDfm, wgt = TRUE)
}
}
