% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conserved.R
\name{mirProcess}
\alias{mirProcess}
\title{mirProcess}
\usage{
mirProcess(path = getwd(), species = NULL, target.annot.file = NULL,
  database = "mirbase", raw.file.sep = "", parallelComputing = FALSE,
  clusterType = "FORK")
}
\arguments{
\item{path}{Path to raw files. Default is the system working directory.}

\item{species}{Species code, following the traditional abbreviated naming convention, e.g. "hsa", "mmu".}

\item{target.annot.file}{Annotation file describing filenames and targets, and should be in \code{csv} format.}

\item{database}{MiRNA database, only for miRNA naming conventions. Currently the function only takes "mirbase".}

\item{raw.file.sep}{Raw read count file separators. Default is \code{""\"\"}, i.e. white space.}

\item{parallelComputing}{Wether to use parallel computing or not. Default is \code{TRUE}.}

\item{cluterType}{clusterType Only set when \code{parallelComputing = TRUE}, the type for parallel cluster. Options are \code{"PSOCK"} (all operating systems) and \code{"FORK"} (macOS and Unix-like system only). Default is \code{"PSOCK"}.}
}
\value{
Outputs a \code{mirna_count} with merged read counts from mutliple files, with annotation.
}
\description{
Data pre-processing for miRNA-seq read count files. Also eee \code{\link{mirProcessML}}.
}
\details{
When \code{species} left as \code{NULL}, the function will use all the species detected in the raw read count files.
         The raw count files are usually in \code{.txt} format with "read count" and "mirna" columns. The read count files can be obtained using the shell program \code{mirna_processing}.
}
\examples{
\dontrun{
readcountMerged <- mirProcess()
}
}
